dirmngr.install: remove ldap file
Even with the dependencies installed, autoconf doesn't detect proper
ldap support and builds dirmngr without it, so remove the file to avoid
the deb package building phase to fail (since it'll be missing).

I have no idea why it doesn't properly detect ldap, and have no time to
fuck around with this.