base-passwd/control: add docbook to build deps
I don't know why it's not there in Debian's package control file, maybe
under normal circumstances it's supposed to be installed automatically
with the other dependencies, but it wasn't and the package fails to
build due to openjade not being able to handle the sgml docs. Installing
docbook fixed the issue.